Refactored Student and Professor entities to domain package.
